+++ ExportCategories.lua
@@ -1106,6 +1106,7 @@ DF.ExportCategories = {
         "oorPowerBarAlpha",
         "oorAurasAlpha",
         "oorNameTextAlpha",
+        "oorTextAlpha",
         "oorIconsAlpha",
         "oorDispelOverlayAlpha",
         "oorDefensiveIconAlpha",